MeshImporter/XFormImporter refactor
Sanitizing the USD Meshes and XForms to Unity standards is now done by the sample itself right after deserialization. This has the
benefit to run in jobs and to simplify the MeshImporter code greatly.
Sanitization code takes care of:
    - xform conversion
    - triangulation
    - de-indexation of attributes
    - attribute scope conversion

The SkeletonImporter has been changed and the JointWeights and JointIndices are now deserialized/loaded by the MeshImporter to gain a whole extra mesh deserialization/sanitization.